remove deprecated Beam
