+++ onmt/translate/__init__.py
@@ -1,15 +1,14 @@
 """ Modules for translation """
 from onmt.translate.translator import Translator
 from onmt.translate.translation import Translation, TranslationBuilder
-from onmt.translate.beam import Beam, GNMTGlobalScorer
-from onmt.translate.beam_search import BeamSearch
+from onmt.translate.beam_search import BeamSearch, GNMTGlobalScorer
 from onmt.translate.decode_strategy import DecodeStrategy
 from onmt.translate.greedy_search import GreedySearch
 from onmt.translate.penalties import PenaltyBuilder
 from onmt.translate.translation_server import TranslationServer, \
     ServerModelError
 
-__all__ = ['Translator', 'Translation', 'Beam', 'BeamSearch',
+__all__ = ['Translator', 'Translation', 'BeamSearch',
            'GNMTGlobalScorer', 'TranslationBuilder',
            'PenaltyBuilder', 'TranslationServer', 'ServerModelError',
            "DecodeStrategy", "GreedySearch"]
